Move line extract function to the migrate package
