+++ migrate/error_line_extract.go
@@ -1,4 +1,4 @@
-package main
+package migrate
 
 import (
 	"fmt"
