+++ migrate/error_line_extract_test.go
@@ -1,4 +1,4 @@
-package main
+package migrate
 
 import (
 	"testing"
